fixed cross sections, renamed tth_hbb to tth
