+++ MEAnalysis/gc/datasets/sig.dat
@@ -1,2 +1,2 @@
-[tth_hbb_13tev]
+[tth_13tev]
 /store/user/jpata/VHBBHeppyV10/TTbarH_M-125_13TeV_amcatnlo-pythia8-tauola/VHBB_HEPPY_V10_TTbarH_M-125_13TeV_amcatnlo-pythia8-tauola__Phys14DR-PU20bx25_tsg_PHYS14_25_V1-v2/150302_164515/0000/tree_1.root = 198284
